+++ CHANGELOG.md
@@ -143,3 +143,4 @@
 - 2023-10-23: Test deals, qualifier
 - 2023-10-23: Serde deals, qualifier, gp, scheduler
 - 2023-10-24: Pagination deals, session, gp
+- 2023-10-25: Serde deals, session